further refactoring; supporting objective function of different cost definition
